[#25] refactor : 함수 및 클래스 네이밍 변경
