+++ coupon-platform-domain/src/main/kotlin/coupon_platform/domain/issued_coupon/repository/IssuedCouponStore.kt
@@ -3,5 +3,5 @@ package coupon_platform.domain.issued_coupon.repository
 import coupon_platform.domain.issued_coupon.entitiy.IssuedCoupon
 
 interface IssuedCouponStore {
-    fun issueCoupon(issuedCoupon: IssuedCoupon): IssuedCoupon
+    fun saveIssuedCoupon(issuedCoupon: IssuedCoupon): IssuedCoupon
 }
